+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - `IMPROVED` #379 处理 Selected `$.fn.selected` 与 [jQuery Form](https://github.com/malsup/form) 冲突问题， Thx @Ahoo-Wang；
 - `IMPROVED` #367 尝试修复 Modal 遮罩层关闭回调函数被触发多次问题；
+- `UPDATE` 升级 FullScreen 至 `2.0.0`，移除 `.onchange()`、`.onerror()` 方法。
 
 ## Amaze UI 2.1.0-beta1